+++ pyproject.toml
@@ -57,7 +57,7 @@ langchain = ["langchain-core>=1.2.5,<2"]
 # extra above would push LangGraph onto every `guard_tool` user.
 #
 # `langgraph` is named even though `langchain` depends on it, because
-# `middleware.py` imports `langgraph.prebuilt.tool_node` and `langgraph.types`
+# `_middleware.py` imports `langgraph.prebuilt.tool_node` and `langgraph.types`
 # directly. The floors are the versions the middleware actually imports at:
 # below langgraph 1.2 `tool_node` fails on `langgraph.runtime.ExecutionInfo`,
 # and langchain below 1.3 resolves to one of those. The core floor matches the
